Fixing numpy usage in tests

+++ tests/unit/qm/liouvillespace/test_redfield.py
@@ -27,6 +27,7 @@ from quantarhei.qm import SystemBathInteraction
 from quantarhei.qm import ReducedDensityMatrixPropagator
 from quantarhei.qm import ReducedDensityMatrix
 from quantarhei import energy_units
+from quantarhei import REAL
 
 class TestRedfield(unittest.TestCase):
     """Tests for the RedfieldRelaxationTensor class
@@ -54,8 +55,8 @@ class TestRedfield(unittest.TestCase):
         cm2 = CorrelationFunctionMatrix(time,2,1)
         cm2.set_correlation_function(cf2,[(0,0),(1,1)])  
           
-        K11 = numpy.array([[1.0, 0.0],[0.0, 0.0]],dtype=numpy.float)
-        K21 = numpy.array([[1.0, 0.0],[0.0, 0.0]],dtype=numpy.float)
+        K11 = numpy.array([[1.0, 0.0],[0.0, 0.0]],dtype=REAL)
+        K21 = numpy.array([[1.0, 0.0],[0.0, 0.0]],dtype=REAL)
         K12 = K11.copy()
         K22 = K21.copy()
             
@@ -90,8 +91,8 @@ class TestRedfield(unittest.TestCase):
         matrix = True
         
         dim = self.H1.dim
-        KT = numpy.zeros((dim,dim), dtype=numpy.float64)
-        KM = numpy.zeros((dim,dim), dtype=numpy.float64)
+        KT = numpy.zeros((dim,dim), dtype=REAL)
+        KM = numpy.zeros((dim,dim), dtype=REAL)
         
         if tensor:
             #print(self.H1)
